backend/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-ranged-response==0.2.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-simple-captcha==0.6.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
redis==7.1.0
sqlparse==0.5.5
